nest delete 엔드포인트 여러 id 삭제하는 것으로 변경

+++ src/nest/src/translation/translation.service.ts
@@ -40,7 +40,16 @@ export class TranslationService {
     await this.translationRepository.editTranslation(params);
   }
 
-  async deleteTranslation({ id }: { id: UUID }) {
-    await this.translationRepository.deleteTranslation({ id });
+  async deleteTranslations({ ids }: { ids: UUID[] }) {
+    for (const id of ids) {
+      if (!this.isUUID(id)) throw new Error('Invalid UUID');
+      await this.translationRepository.deleteTranslation({ id });
+    }
+  }
+
+  private isUUID(id: string) {
+    const uuidRegex =
+      /^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$/i;
+    return uuidRegex.test(id);
   }
 }
